+++ Sources/MapboxNavigationCore/Map/Style/NavigationMapStyleManager.swift
@@ -34,6 +34,7 @@ struct MapStyleConfig: Equatable {
     var routeAnnnotationCaptionTextFont: UIFont
 
     var routeLineTracksTraversal: Bool
+    var routeLineWidthMultiplier: Double
     var isRestrictedAreaEnabled: Bool
     var showsTrafficOnRouteLine: Bool
     var showsAlternatives: Bool