+++ Sources/_MapboxNavigationTestHelpers/Core/Map/Style/MapStyleConfigMock.swift
@@ -5,6 +5,7 @@ import UIKit
 extension MapStyleConfig {
     public static func mock(
         routeLineTracksTraversal: Bool = true,
+        routeLineWidthMultiplier: Double = 1.0,
         isRestrictedAreaEnabled: Bool = true,
         showsTrafficOnRouteLine: Bool = true,
         showsAlternatives: Bool = true,
@@ -41,6 +42,7 @@ extension MapStyleConfig {
             routeAnnotationTextFont: .systemFont(ofSize: 18, weight: .semibold),
             routeAnnnotationCaptionTextFont: .systemFont(ofSize: 16, weight: .regular),
             routeLineTracksTraversal: routeLineTracksTraversal,
+            routeLineWidthMultiplier: routeLineWidthMultiplier,
             isRestrictedAreaEnabled: isRestrictedAreaEnabled,
             showsTrafficOnRouteLine: showsTrafficOnRouteLine,
             showsAlternatives: showsAlternatives,